+++ src/httphandler/BaseHttpHandler.java
@@ -2,10 +2,6 @@ package httphandler;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
-import com.google.gson.TypeAdapter;
-import com.google.gson.stream.JsonReader;
-import com.google.gson.stream.JsonToken;
-import com.google.gson.stream.JsonWriter;
 import com.sun.net.httpserver.HttpExchange;
 import com.sun.net.httpserver.HttpHandler;
 import general.Const;
@@ -17,46 +13,6 @@ import java.io.OutputStream;
 import java.nio.charset.StandardCharsets;
 import java.time.Duration;
 import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-import java.util.Objects;
-
-class DurationAdapter extends TypeAdapter<Duration> {
-
-    @Override
-    public void write(final JsonWriter jsonWriter, final Duration duration) throws IOException {
-        jsonWriter.value(Objects.isNull(duration) ? null : duration.toMinutes());
-    }
-
-    @Override
-    public Duration read(final JsonReader jsonReader) throws IOException {
-        if (jsonReader.peek() == JsonToken.NULL) {
-            jsonReader.nextNull();
-            return null;
-        } else {
-            return Duration.ofMinutes(Integer.parseInt(jsonReader.nextString()));
-        }
-    }
-}
-
-class LocalDateTimeAdapter extends TypeAdapter<LocalDateTime> {
-
-    private static final DateTimeFormatter dtf = DateTimeFormatter.ISO_LOCAL_DATE_TIME;
-
-    @Override
-    public void write(final JsonWriter jsonWriter, final LocalDateTime localDateTime) throws IOException {
-        jsonWriter.value(Objects.isNull(localDateTime) ? null : localDateTime.format(dtf));
-    }
-
-    @Override
-    public LocalDateTime read(final JsonReader jsonReader) throws IOException {
-        if (jsonReader.peek() == JsonToken.NULL) {
-            jsonReader.nextNull();
-            return null;
-        } else {
-            return LocalDateTime.parse(jsonReader.nextString(), dtf);
-        }
-    }
-}
 
 public class BaseHttpHandler implements HttpHandler {
 
@@ -69,7 +25,7 @@ public class BaseHttpHandler implements HttpHandler {
     protected InputStream inputStream;
     protected String jsonBody;
 
-    protected final static Gson gson = new GsonBuilder()
+    protected static final Gson gson = new GsonBuilder()
             .serializeNulls()
             .setPrettyPrinting()
             .registerTypeAdapter(Duration.class, new DurationAdapter())
