+++ src/httphandler/DurationAdapter.java
@@ -1,0 +1,28 @@
+package httphandler;
+
+import com.google.gson.TypeAdapter;
+import com.google.gson.stream.JsonReader;
+import com.google.gson.stream.JsonToken;
+import com.google.gson.stream.JsonWriter;
+
+import java.io.IOException;
+import java.time.Duration;
+import java.util.Objects;
+
+class DurationAdapter extends TypeAdapter<Duration> {
+
+    @Override
+    public void write(final JsonWriter jsonWriter, final Duration duration) throws IOException {
+        jsonWriter.value(Objects.isNull(duration) ? null : duration.toMinutes());
+    }
+
+    @Override
+    public Duration read(final JsonReader jsonReader) throws IOException {
+        if (jsonReader.peek() == JsonToken.NULL) {
+            jsonReader.nextNull();
+            return null;
+        } else {
+            return Duration.ofMinutes(Integer.parseInt(jsonReader.nextString()));
+        }
+    }
+}
