+++ src/httphandler/LocalDateTimeAdapter.java
@@ -1,0 +1,31 @@
+package httphandler;
+
+import com.google.gson.TypeAdapter;
+import com.google.gson.stream.JsonReader;
+import com.google.gson.stream.JsonToken;
+import com.google.gson.stream.JsonWriter;
+
+import java.io.IOException;
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+import java.util.Objects;
+
+class LocalDateTimeAdapter extends TypeAdapter<LocalDateTime> {
+
+    private static final DateTimeFormatter dtf = DateTimeFormatter.ISO_LOCAL_DATE_TIME;
+
+    @Override
+    public void write(final JsonWriter jsonWriter, final LocalDateTime localDateTime) throws IOException {
+        jsonWriter.value(Objects.isNull(localDateTime) ? null : localDateTime.format(dtf));
+    }
+
+    @Override
+    public LocalDateTime read(final JsonReader jsonReader) throws IOException {
+        if (jsonReader.peek() == JsonToken.NULL) {
+            jsonReader.nextNull();
+            return null;
+        } else {
+            return LocalDateTime.parse(jsonReader.nextString(), dtf);
+        }
+    }
+}
